# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Education Management API List",
        "version": "1.0"
    },
    "paths": {
...
        "/api/topics": {
            "get": {
                "tags": [
                    "Topics"
                ],
                "summary": "Get Topics list",
                "description": "Returns Topics list",
                "operationId": "getTopicsList",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Topics"
                ],
                "summary": "Create Topic",
                "description": "",
                "operationId": "createNewTopics",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Name of topic",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Danh muc 1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        },
        "/api/topics/{id}": {
            "get": {
                "tags": [
                    "Topics"
                ],
                "summary": "Get details of Topic",
                "description": "Returns Topic details",
                "operationId": "getTopicDetails",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "topic_id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Topics"
                ],
                "summary": "Create Topic",
                "description": "",
                "operationId": "createUpdateTopics",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "id of topic",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Topics"
                ],
                "summary": "Delete topics",
                "description": "Returns status delete",
                "operationId": "deleteTopic",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "topic_id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        },
...
    }
}